requirements.txt
aiogram>=3.4.0
openai>=1.12.0
yt-dlp>=2024.1.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0
